ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Machine Learning Libraries
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
prophet==1.1.5
tensorflow==2.15.0  # Optional: for deep learning models

# Data Analysis
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# HTTP Client (for calling Node.js backend)
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-multipart==0.0.6
